requirements.txt
beautifulsoup4
Flask
selenium
google-auth
google-api-python-client
pywebview
